requirements.txt
asgiref==3.8.1
Django==5.1.2
django-allauth==65.1.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.8.2
